+++ src/main/java/io/bootique/liquibase/command/ChangelogSyncCommand.java
@@ -5,12 +5,15 @@ import com.google.inject.Provider;
 import io.bootique.cli.Cli;
 import io.bootique.command.CommandOutcome;
 import io.bootique.command.CommandWithMetadata;
+import io.bootique.liquibase.LiquibaseModule;
 import io.bootique.liquibase.LiquibaseRunner;
 import io.bootique.meta.application.CommandMetadata;
 import liquibase.Contexts;
 import liquibase.LabelExpression;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.List;
 
 /**
  * @since 0.11
@@ -37,7 +40,9 @@ public class ChangelogSyncCommand extends CommandWithMetadata {
 
         return runnerProvider.get().runWithLiquibase(lb -> {
             try {
-                lb.changeLogSync(new Contexts(), new LabelExpression());
+                List<String> options = cli.optionStrings(LiquibaseModule.CONTEXT_OPTION);
+                lb.changeLogSync(options == null ? new Contexts() : new Contexts(options.toArray(new String[options.size()])), new LabelExpression());
+
                 return CommandOutcome.succeeded();
             } catch (Exception e) {
                 return CommandOutcome.failed(1, e);