+++ src/main/java/io/bootique/liquibase/command/UpdateCommand.java
@@ -2,15 +2,18 @@ package io.bootique.liquibase.command;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-import io.bootique.meta.application.CommandMetadata;
 import io.bootique.cli.Cli;
 import io.bootique.command.CommandOutcome;
 import io.bootique.command.CommandWithMetadata;
+import io.bootique.liquibase.LiquibaseModule;
 import io.bootique.liquibase.LiquibaseRunner;
+import io.bootique.meta.application.CommandMetadata;
 import liquibase.Contexts;
 import liquibase.LabelExpression;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.List;
 
 
 public class UpdateCommand extends CommandWithMetadata {
@@ -36,7 +39,9 @@ public class UpdateCommand extends CommandWithMetadata {
 
 		return runnerProvider.get().runWithLiquibase(lb -> {
 			try {
-				lb.update(new Contexts(), new LabelExpression());
+				List<String> options = cli.optionStrings(LiquibaseModule.CONTEXT_OPTION);
+				lb.update(options == null ? new Contexts() : new Contexts(options.toArray(new String[options.size()])), new LabelExpression());
+
 				return CommandOutcome.succeeded();
 			} catch (Exception e) {
 				return CommandOutcome.failed(1, e);